requirements.txt
python-dotenv~=1.0.1
pydantic==2.8.2
gradio==4.44.1
groq